trying to get the traces to be visible in the analyser view

+++ AUDIO_2/Inc/phasor.h
@@ -5,7 +5,7 @@ typedef struct _tPhasor {
 	float phase; // [0.0, 1.0)
 	float inc;
 	int(*freq)(struct _tPhasor *self, float freq);
-	uint16_t(*step)(struct _tPhasor *self);
+	float(*step)(struct _tPhasor *self);
 	//uint16_t(*samp)(struct _tPhasor *self);
 } tPhasor;
